requirements.txt
pychalk
aiohttp
miniupnpc
merge3